requirements.txt
requests==2.31.0
openpyxl==3.1.0
pandas==2.1.4
pydantic==2.5.0
python-dotenv==1.0.0

# Schwab API (replaces Finnhub)
schwab-py>=1.5.0
httpx>=0.27.0        # schwab-py uses httpx under the hood
